Move the ReadOnlyMemory overload to a partial class outside the common NetEventSource
