+++ src/Common/src/System/Net/Logging/NetEventSource.Common.cs
@@ -287,23 +287,7 @@ namespace System.Net
         {
             DumpBuffer(thisOrContextObject, buffer, 0, buffer.Length, memberName);
         }
-
-        /// <summary>Logs the contents of a buffer </summary>
-        /// <param name="thisOrContextObject">`this`, or another object that serves to provide context for the operation.</param>
-        /// <param name="buffer">The buffer to be logged.</param>
-        /// <param name="memberName">The calling member.</param>
-        [NonEvent]
-        public static void DumpBuffer(object thisOrContextObject, ReadOnlyMemory<byte> buffer, [CallerMemberName] string memberName = null)
-        {
-            if (IsEnabled)
-            {
-                int count = Math.Min(buffer.Length, MaxDumpSize);
-
-                byte[] slice = buffer.Slice(0, count).ToArray();
-                Log.DumpBuffer(IdOf(thisOrContextObject), memberName, slice);
-            }
-        }
-
+                
         /// <summary>Logs the contents of a buffer.</summary>
         /// <param name="thisOrContextObject">`this`, or another object that serves to provide context for the operation.</param>
         /// <param name="buffer">The buffer to be logged.</param>
